EY-2118: Stokkar om på flyten sånn at vi først køyrer inn alt som opplysninger fra PDL av seg sjølv, og etter det fyllar vi på med persongalleriet. Med den rekkjefølgja unngår vi å stoppe i rejectKey i opplysninger-fra-pdl
