+++ apps/etterlatte-oppdater-behandling/src/main/kotlin/FortsettMigreringsflyten.kt
@@ -1,13 +1,17 @@
 package no.nav.etterlatte
 
+import com.fasterxml.jackson.module.kotlin.treeToValue
 import no.nav.etterlatte.libs.common.behandling.SakType
 import no.nav.etterlatte.libs.common.grunnlag.opplysningstyper.Opplysningstype
 import no.nav.etterlatte.libs.common.logging.withLogContext
+import no.nav.etterlatte.libs.common.objectMapper
 import no.nav.etterlatte.libs.common.person.PersonRolle
 import no.nav.etterlatte.libs.common.rapidsandrivers.BEHOV_NAME_KEY
 import no.nav.etterlatte.libs.common.rapidsandrivers.SAK_TYPE_KEY
 import no.nav.etterlatte.libs.common.rapidsandrivers.correlationId
 import no.nav.etterlatte.libs.common.rapidsandrivers.eventName
+import no.nav.etterlatte.rapidsandrivers.migrering.FNR_KEY
+import no.nav.etterlatte.rapidsandrivers.migrering.MigreringRequest
 import no.nav.etterlatte.rapidsandrivers.migrering.Migreringshendelser
 import no.nav.etterlatte.rapidsandrivers.migrering.ROLLE_KEY
 import no.nav.helse.rapids_rivers.JsonMessage
@@ -16,10 +20,15 @@ import no.nav.helse.rapids_rivers.RapidsConnection
 import no.nav.helse.rapids_rivers.River
 import org.slf4j.LoggerFactory
 import rapidsandrivers.BEHANDLING_ID_KEY
-import rapidsandrivers.GRUNNLAG_OPPDATERT
+import rapidsandrivers.HENDELSE_DATA_KEY
+import rapidsandrivers.behandlingId
+import rapidsandrivers.sakId
 import rapidsandrivers.withFeilhaandtering
 
-internal class FortsettMigreringsflyten(rapidsConnection: RapidsConnection) :
+internal class FortsettMigreringsflyten(
+    rapidsConnection: RapidsConnection,
+    private val behandlinger: BehandlingService
+) :
     River.PacketListener {
 
     private val logger = LoggerFactory.getLogger(this::class.java)
@@ -27,19 +36,28 @@ internal class FortsettMigreringsflyten(rapidsConnection: RapidsConnection) :
     init {
         logger.info("initierer rapid for migreringshendelser")
         River(rapidsConnection).apply {
-            eventName(GRUNNLAG_OPPDATERT)
+
+            eventName(Migreringshendelser.MIGRER_SAK)
 
             correlationId()
-            validate { it.rejectValue(BEHOV_NAME_KEY, Opplysningstype.MIGRERING.name) }
-            validate { it.requireKey(BEHANDLING_ID_KEY) }
-            validate { it.requireKey("opplysning") }
+            validate { it.rejectKey(BEHANDLING_ID_KEY) }
+            validate { it.requireKey(HENDELSE_DATA_KEY) }
+            validate { it.requireKey(FNR_KEY) }
+            validate { it.rejectKey("opplysning") }
         }.register(this)
     }
 
     override fun onPacket(packet: JsonMessage, context: MessageContext) {
         withLogContext(packet.correlationId) {
-            withFeilhaandtering(packet, context, Migreringshendelser.START_MIGRERING) {
+            withFeilhaandtering(packet, context, Migreringshendelser.MIGRER_SAK) {
                 logger.info("Mottatt migreringshendelse")
+
+                val hendelse: MigreringRequest = objectMapper.treeToValue(packet[HENDELSE_DATA_KEY])
+                val (behandlingId, sakId) = behandlinger.migrer(hendelse)
+
+                packet.behandlingId = behandlingId
+                packet.sakId = sakId
+
                 packet[BEHOV_NAME_KEY] = Opplysningstype.MIGRERING
 
                 packet[SAK_TYPE_KEY] = SakType.BARNEPENSJON
