+++ apps/etterlatte-oppdater-behandling/src/main/kotlin/MigreringHendelser.kt
@@ -1,29 +1,27 @@
 package no.nav.etterlatte
 
-import com.fasterxml.jackson.module.kotlin.treeToValue
 import no.nav.etterlatte.libs.common.behandling.Persongalleri
 import no.nav.etterlatte.libs.common.grunnlag.Grunnlagsopplysning
 import no.nav.etterlatte.libs.common.grunnlag.opplysningstyper.Opplysningstype
 import no.nav.etterlatte.libs.common.logging.withLogContext
 import no.nav.etterlatte.libs.common.objectMapper
+import no.nav.etterlatte.libs.common.rapidsandrivers.BEHOV_NAME_KEY
 import no.nav.etterlatte.libs.common.rapidsandrivers.correlationId
 import no.nav.etterlatte.libs.common.rapidsandrivers.eventName
-import no.nav.etterlatte.rapidsandrivers.migrering.FNR_KEY
 import no.nav.etterlatte.rapidsandrivers.migrering.MigreringRequest
 import no.nav.etterlatte.rapidsandrivers.migrering.Migreringshendelser.MIGRER_SAK
+import no.nav.etterlatte.rapidsandrivers.migrering.request
 import no.nav.helse.rapids_rivers.JsonMessage
 import no.nav.helse.rapids_rivers.MessageContext
 import no.nav.helse.rapids_rivers.RapidsConnection
 import no.nav.helse.rapids_rivers.River
 import org.slf4j.LoggerFactory
 import rapidsandrivers.BEHANDLING_ID_KEY
-import rapidsandrivers.HENDELSE_DATA_KEY
-import rapidsandrivers.behandlingId
-import rapidsandrivers.sakId
+import rapidsandrivers.GRUNNLAG_OPPDATERT
 import rapidsandrivers.withFeilhaandtering
 import java.util.*
 
-internal class MigreringHendelser(rapidsConnection: RapidsConnection, private val behandlinger: BehandlingService) :
+internal class MigreringHendelser(rapidsConnection: RapidsConnection) :
     River.PacketListener {
 
     private val logger = LoggerFactory.getLogger(this::class.java)
@@ -31,13 +29,12 @@ internal class MigreringHendelser(rapidsConnection: RapidsConnection, private va
     init {
         logger.info("initierer rapid for migreringshendelser")
         River(rapidsConnection).apply {
-            eventName(MIGRER_SAK)
+            eventName(GRUNNLAG_OPPDATERT)
 
             correlationId()
-            validate { it.rejectKey(BEHANDLING_ID_KEY) }
-            validate { it.requireKey(HENDELSE_DATA_KEY) }
-            validate { it.requireKey(FNR_KEY) }
-            validate { it.rejectKey("opplysning") }
+            validate { it.requireValue(BEHOV_NAME_KEY, Opplysningstype.MIGRERING.name) }
+            validate { it.requireKey(BEHANDLING_ID_KEY) }
+            validate { it.requireKey("opplysning") }
         }.register(this)
     }
 
@@ -45,14 +42,10 @@ internal class MigreringHendelser(rapidsConnection: RapidsConnection, private va
         withLogContext(packet.correlationId) {
             withFeilhaandtering(packet, context, MIGRER_SAK) {
                 logger.info("Mottatt migreringshendelse, klar til å opprette persongalleri")
-
-                val hendelse: MigreringRequest = objectMapper.treeToValue(packet[HENDELSE_DATA_KEY])
-                val (behandlingId, sakId) = behandlinger.migrer(hendelse)
-
-                packet.behandlingId = behandlingId
-                packet.sakId = sakId
-                packet["opplysning"] = tilOpplysning(hendelse.persongalleri)
+                val request = objectMapper.readValue(packet.request, MigreringRequest::class.java)
+                packet["opplysning"] = tilOpplysning(request.persongalleri)
                 packet.eventName = "OPPLYSNING:NY"
+                packet["fullstendig"] = true
                 context.publish(packet.toJson())
                 logger.info("Publiserte persongalleri")
             }
